refactor(Element): Remove 'fake' HOC
Flow was having some trouble typing the Element function correctly,
which is probably due to it looking like a HOC but not really acting one
since it doesn't take in a React Component as a parameter

To fix this, I refactoring that function to be a simple stateless
function that passes its parameters as props.

I've also adjusted the tests appropriately.