Use new changes in demo
In a world with no types, there are **no integration** changes needed.

In a hypothetical world with types, this diff corresponds to the changes
our users would have to deal with. Note how they have to check
`if (this.props.stripe)` even though we can statically guarantee it will
always be defined.