Add new async demo and tests
The async demo is copy/paste from the main demo.
The only differences are *deletions*. This means no new code to review.

It works by crafting and injecting a `<script>` tag into the page.

It's served from http://localhost:8080/async/.
In addition, it's "server-side" rendered in our tests (smoke test).